created basic division
